requirements.txt
Flask==3.0.0
tensorflow==2.21.0
Pillow
pillow-heif
Werkzeug==3.0.1
